fix(indent): Remove indent rule
Indentation rules in TypeScript together with eslint is extremely
complex. Mixing with settings makes this even worse.

Disabling the eslint indentation rule and letting prettier decide what
to do instead.

My specific problem was that eslint and prettier was not consistent
about the indentation of long lines of type-aliases as return-types.
Example of how prettier wants it:

	function example (): Example<
		ThisIsALongTypeT,
		ThisIsAnotherOneT,
		EtcT
	> {
		// ...
	}

Example how eslint wants it (with the indentation rule):

	function example (): Example<
		ThisIsALongTypeT,
		ThisIsAnotherOneT,
		EtcT
		> {
		// ...
	}

Related links:

 - https://github.com/typescript-eslint/typescript-eslint/issues/1824
 - https://github.com/sveltejs/eslint-config/pull/1

After removing this line, prettier does a nice job of indenting it
correctly, and eslint doesn't complain.